JobService: add "directory creation" job
Adds a "mkdir" job in masquerade (which can be called by substratum).
It allows creating folders in directories otherwise unaccessible without root
(but it still has its limitations, because it uses the "system" user).

It also allows creating the "data/system/theme" folder, so this commit:
http://review.projektsubstratum.com/#/c/280/ should not be needed.

a combined work with KreAch3R

Change-Id: I49e79dc459a9348493389c009c79ed99bfec9ab6